refactor: split api.ts into domain-specific modules
